requirements.txt
accelerate==1.6.0
clip
datasets==3.2.0
diffusers==0.33.1
matplotlib==3.10.1
nudenet==3.4.2
numpy==2.2.4
openai==0.28.0
openai-clip==1.0.1
Pillow==12.1.1
scikit-learn==1.6.1
seaborn==0.13.2
torch==2.6.0
torch-fidelity==0.3.0
torchmetrics==1.7.1
torchvision==0.21.0
transformers==4.51.3
wandb==0.25.1
